requirements.txt
discord.py == 2.4.0
discord-ext-voice-recv
python-dotenv
pydub
websockets
audioop-lts
openai
google-genai
